Business Plan for Importing and Exporting boat engines
This business operates in the import and export industry of boat engines.
The purchase of boat engines from countries where cost of manufacturing is lower (like China) and export to countries where purchasing power is higher (Local market, Europe, Australia, and New Zealand). The goals of the business is to become a leader in the boat engine import & export industry and provide the best quality boat engines to customers in order create value and the best customer experience.
In this presentation, I would like to present product information, regulation of import and export, supply & demand analysis, competition, target markets, logistics, prices, promotional activities, financial aspects and sources of product information.